+++ app/src/main/java/com/larryhsiao/nyx/ViewModelFactory.kt
@@ -3,6 +3,7 @@ package com.larryhsiao.nyx
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.larryhsiao.nyx.jot.JotViewModel
+import com.larryhsiao.nyx.jot.JotsCalendarViewModel
 import com.larryhsiao.nyx.jot.JotsViewModel
 import com.larryhsiao.nyx.old.sync.LocalFileSync
 
@@ -12,11 +13,14 @@ import com.larryhsiao.nyx.old.sync.LocalFileSync
 class ViewModelFactory(private val app: JotApplication) : ViewModelProvider.Factory {
     override fun <T : ViewModel?> create(modelClass: Class<T>): T {
         return when {
-            modelClass.isAssignableFrom(JotsViewModel::class.java) -> {
-                JotsViewModel(app) as T
+            modelClass.isAssignableFrom(JotsCalendarViewModel::class.java) -> {
+                JotsCalendarViewModel(app) as T
             }
             modelClass.isAssignableFrom(JotViewModel::class.java) -> {
                 JotViewModel(app.db, LocalFileSync(app, app.db)) as T
+            }
+            modelClass.isAssignableFrom(JotsViewModel::class.java) -> {
+                JotsViewModel(app) as T
             }
             else -> {
                 modelClass.getConstructor().newInstance()
