+++ app/src/main/res/layout/fragment_jots_calendar.xml
@@ -1,0 +1,136 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    xmlns:tools="http://schemas.android.com/tools"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <TextView
+        android:layout_width="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        android:id="@+id/jots_month_textView"
+        android:padding="8dp"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        tools:text="YYYY-MM-DD"
+        android:background="?attr/selectableItemBackground"
+        android:textAppearance="@style/TextAppearance.AppCompat.Title"
+        android:layout_height="wrap_content" />
+
+    <com.haibin.calendarview.CalendarLayout
+        android:layout_width="match_parent"
+        app:default_status="shrink"
+        app:layout_constraintTop_toBottomOf="@id/jots_month_textView"
+        app:layout_constraintBottom_toBottomOf="parent"
+        android:layout_height="0dp">
+
+        <com.haibin.calendarview.CalendarView
+            android:layout_width="match_parent"
+            android:id="@+id/jots_calendarView"
+            android:layout_height="wrap_content" />
+
+        <androidx.constraintlayout.widget.ConstraintLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent">
+
+            <ImageView
+                android:id="@+id/jot_list_map_switcher_imageView"
+                android:layout_width="wrap_content"
+                app:layout_constraintEnd_toEndOf="parent"
+                android:src="@drawable/ic_map"
+                android:background="?attr/selectableItemBackgroundBorderless"
+                android:layout_marginEnd="8dp"
+                app:layout_constraintTop_toTopOf="@id/jots_day_textView"
+                app:layout_constraintBottom_toBottomOf="@id/jots_day_textView"
+                android:padding="8dp"
+                android:layout_height="wrap_content" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:id="@+id/jots_day_textView"
+                tools:text="MM-DD"
+                android:layout_margin="16dp"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintStart_toStartOf="parent"
+                android:textAppearance="@style/TextAppearance.MaterialComponents.Subtitle2"
+                android:layout_height="wrap_content" />
+
+            <ImageView
+                android:layout_width="wrap_content"
+                app:layout_constraintStart_toEndOf="@id/jots_day_textView"
+                android:id="@+id/jots_newJot_imageView"
+                android:layout_marginStart="8dp"
+                android:padding="8dp"
+                app:layout_constraintTop_toTopOf="@id/jots_day_textView"
+                app:layout_constraintBottom_toBottomOf="@id/jots_day_textView"
+                android:gravity="center"
+                android:textAppearance="@style/TextAppearance.AppCompat.Body1"
+                android:background="?attr/selectableItemBackgroundBorderless"
+                android:layout_height="wrap_content"
+                android:src="@drawable/ic_plus" />
+
+            <ImageView
+                android:layout_width="wrap_content"
+                app:layout_constraintStart_toEndOf="@id/jots_newJot_imageView"
+                android:id="@+id/jots_newJotByImage_imageView"
+                android:padding="8dp"
+                app:layout_constraintTop_toTopOf="@id/jots_day_textView"
+                app:layout_constraintBottom_toBottomOf="@id/jots_day_textView"
+                android:gravity="center"
+                android:textAppearance="@style/TextAppearance.AppCompat.Body1"
+                android:background="?attr/selectableItemBackgroundBorderless"
+                android:layout_height="wrap_content"
+                android:src="@drawable/ic_image_plus" />
+
+            <androidx.recyclerview.widget.RecyclerView
+                android:layout_width="match_parent"
+                app:layout_constraintTop_toBottomOf="@id/jots_day_textView"
+                android:id="@+id/jots_recyclerView"
+                android:layout_marginTop="16dp"
+                app:layout_constraintBottom_toBottomOf="parent"
+                app:layoutManager="androidx.recyclerview.widget.LinearLayoutManager"
+                tools:listitem="@android:layout/simple_list_item_1"
+                android:layout_height="0dp" />
+
+            <FrameLayout
+                android:id="@+id/jots_map"
+                android:layout_marginTop="8dp"
+                android:minHeight="100dp"
+                android:minWidth="100dp"
+                app:layout_constraintTop_toBottomOf="@id/jots_day_textView"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:layout_width="match_parent"
+                android:layout_height="0dp" />
+
+            <TextView
+                android:layout_width="wrap_content"
+                android:layout_gravity="center"
+                android:padding="16dp"
+                android:visibility="gone"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"
+                tools:visibility="visible"
+                android:id="@+id/jots_blackhole_textView"
+                android:gravity="center"
+                android:drawablePadding="8dp"
+                android:textSize="12sp"
+                android:textAppearance="@style/TextAppearance.AppCompat.Caption"
+                android:text="@string/Black_hole_destroys_everything_nearby"
+                android:layout_height="wrap_content"
+                app:drawableTopCompat="@drawable/ic_black_hole" />
+
+            <ProgressBar
+                android:layout_width="wrap_content"
+                app:layout_constraintStart_toStartOf="parent"
+                app:layout_constraintEnd_toEndOf="parent"
+                app:layout_constraintTop_toTopOf="parent"
+                app:layout_constraintBottom_toBottomOf="parent"
+                android:id="@+id/jots_loadingBar"
+                android:visibility="gone"
+                tools:visibility="visible"
+                android:layout_height="wrap_content" />
+        </androidx.constraintlayout.widget.ConstraintLayout>
+    </com.haibin.calendarview.CalendarLayout>
+</androidx.constraintlayout.widget.ConstraintLayout>
